Added persistent closures.
Normally closures are freed before a function returns control.

When a closure parent function returns a `Function` type it is
assumed that the returned function relies on the closure and
so it is not freed. In the future when I get around to writing
the garbage collecter it will have to collect these.